nvim: update to v0.12
- Replace `lazy.nvim` with `vim.pack` for plugin management
- Remove unused configs
