+++ ghostty/config
@@ -15,11 +15,13 @@ font-size = 13
 # Switch 'open_config' from 'Ctrl+, to Ctrl+Alt+,'
 keybind = ctrl+comma=unbind
 keybind = ctrl+alt+comma=open_config
+# Paste from clipboard
+keybind = ctrl+shift+v=paste_from_clipboard
 # keybind = performable:ctrl+c=copy_to_clipboard
 
 # Others
 mouse-hide-while-typing = true
-copy-on-select = clipboard
+# copy-on-select = clipboard
 app-notifications = no-clipboard-copy
 
 # Append to config (keep this at the EOF)
