--- nvim/lua/plugins/colorizer.lua
@@ -1,7 +1,0 @@
--- colorizer: https://github.com/norcalli/nvim-colorizer.lua
-return {
-    "norcalli/nvim-colorizer.lua",
-    config = function()
-        require("colorizer").setup()
-    end,
-}
