--- nvim/lua/plugins/marks.lua
@@ -1,6 +1,0 @@
--- marks.nvim: https://github.com/chentoast/marks.nvim
-return {
-    "chentoast/marks.nvim",
-    event = "VeryLazy",
-    opts = {},
-}
