--- nvim/lua/plugins/navbar.lua
@@ -1,35 +1,0 @@
--- nvim-navic: https://github.com/SmiteshP/nvim-navic
-return {
-    "SmiteshP/nvim-navic",
-    opts = {
-        icons = {
-            File = " ",
-            Module = " ",
-            Namespace = " ",
-            Package = " ",
-            Class = " ",
-            Method = " ",
-            Property = " ",
-            Field = " ",
-            Constructor = " ",
-            Enum = " ",
-            Interface = " ",
-            Function = " ",
-            Variable = " ",
-            Constant = " ",
-            String = " ",
-            Number = " ",
-            Boolean = " ",
-            Array = " ",
-            Object = " ",
-            Key = " ",
-            Null = " ",
-            EnumMember = " ",
-            Struct = " ",
-            Event = " ",
-            Operator = " ",
-            TypeParameter = " ",
-        },
-        depth_limit = 3,
-    },
-}
