--- nvim/lua/plugins/preview.lua
@@ -1,7 +1,0 @@
--- https://github.com/brianhuster/live-preview.nvim
-return {
-    "brianhuster/live-preview.nvim",
-    dependencies = {
-        "nvim-telescope/telescope.nvim",
-    },
-}
