--- nvim/lua/plugins/todo.lua
@@ -1,7 +1,0 @@
--- Todo comments: https://github.com/folke/todo-comments.nvim
-return {
-    "folke/todo-comments.nvim",
-    version = "*",
-    dependencies = "nvim-lua/plenary.nvim",
-    config = true,
-}
