+++ nvim/plugin/fidget.lua
@@ -1,0 +1,6 @@
+-- LSP progress status
+vim.pack.add({
+    { src = "https://github.com/j-hui/fidget.nvim", name = "fidget" },
+})
+
+require("fidget").setup()
