+++ nvim/plugin/marks.lua
@@ -1,0 +1,6 @@
+-- marks.nvim: https://github.com/chentoast/marks.nvim
+vim.pack.add({
+    { src = "https://github.com/chentoast/marks.nvim", name = "marks" },
+})
+
+require("marks").setup()
