+++ nvim/plugin/navbar.lua
@@ -1,0 +1,36 @@
+-- nvim-navic: https://github.com/SmiteshP/nvim-navic
+vim.pack.add({
+    { src = "https://github.com/SmiteshP/nvim-navic", name = "nvim-navic" }
+})
+
+require("nvim-navic").setup({
+    icons = {
+        File = " ",
+        Module = " ",
+        Namespace = " ",
+        Package = " ",
+        Class = " ",
+        Method = " ",
+        Property = " ",
+        Field = " ",
+        Constructor = " ",
+        Enum = " ",
+        Interface = " ",
+        Function = " ",
+        Variable = " ",
+        Constant = " ",
+        String = " ",
+        Number = " ",
+        Boolean = " ",
+        Array = " ",
+        Object = " ",
+        Key = " ",
+        Null = " ",
+        EnumMember = " ",
+        Struct = " ",
+        Event = " ",
+        Operator = " ",
+        TypeParameter = " ",
+    },
+    depth_limit = 3,
+})
