+++ nvim/plugin/preview.lua
@@ -1,0 +1,4 @@
+-- https://github.com/brianhuster/live-preview.nvim
+vim.pack.add({
+    { src = "https://github.com/brianhuster/live-preview.nvim", name = "live-preview" },
+})
